requirements.txt
fastapi
uvicorn[standard]
python-multipart
pytesseract
Pillow
PyMuPDF
google-generativeai
python-dotenv
jinja2
